+++ tests/block_header/block_header.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Sonic. If not, see <http://www.gnu.org/licenses/>.
 
-package block_header_test
+package block_header
 
 import (
 	"testing"
